requirements.txt
django
python-dotenv
psycopg2-binary
djangorestframework
flake8
drf-yasg
djangorestframework-simplejwt
django-filter
coverage
